% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutliers.R
\name{findOutliers}
\alias{findOutliers}
\title{FindOutliers function
Finds the outliers in data}
\usage{
findOutliers(x, resp = TRUE)
}
\arguments{
\item{x}{data to analyse}

\item{resp}{Tukey coefficient}
}
\description{
FindOutliers function
Finds the outliers in data
}
