% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimla.R
\name{tanimla}
\alias{tanimla}
\title{Tanimla function
Summarizes the data
DescribesBy data for each factor}
\usage{
tanimla(x, resp, fac)
}
\arguments{
\item{x}{data frame to analyse}

\item{resp}{response variable}

\item{fac}{list of factors}
}
\description{
Tanimla function
Summarizes the data
DescribesBy data for each factor
}
\examples{
tanimla(x, resp, fac)
}
